+++ bw.workflow.deletecompletecriminalrecord/src/main/java/sa/gov/nic/bio/bw/workflow/deletecompletecriminalrecord/utils/DeleteCompleteCriminalRecordErrorCodes.java
@@ -1,0 +1,11 @@
+package sa.gov.nic.bio.bw.workflow.deletecompletecriminalrecord.utils;
+
+public enum DeleteCompleteCriminalRecordErrorCodes
+{
+	C015_00001;
+	
+	public final String getCode()
+	{
+		return name().replace("_", "-");
+	}
+}